refactor: section section 패키지로 분리
