+++ src/main/java/subway/line/Line.java
@@ -1,9 +1,7 @@
 package subway.line;
 
-import subway.station.Station;
-import subway.station.StationDao;
+import subway.section.Section;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.LinkedList;
 import java.util.List;
