--- src/main/java/subway/line/SectionRequest.java
@@ -1,28 +1,0 @@
-package subway.line;
-
-public class SectionRequest {
-    private Long upStationId;
-    private Long downStationId;
-    private int distance;
-
-    public SectionRequest() {
-    }
-
-    public SectionRequest(Long upStationId, Long downStationId, int distance) {
-        this.upStationId = upStationId;
-        this.downStationId = downStationId;
-        this.distance = distance;
-    }
-
-    public Long getUpStationId() {
-        return upStationId;
-    }
-
-    public Long getDownStationId() {
-        return downStationId;
-    }
-
-    public int getDistance() {
-        return distance;
-    }
-}
