+++ src/main/java/subway/section/SectionRequest.java
@@ -1,12 +1,14 @@
-package subway.line;
+package subway.section;
 
-public class SectionResponse {
-
+public class SectionRequest {
     private Long upStationId;
     private Long downStationId;
     private int distance;
 
-    public SectionResponse(Long upStationId, Long downStationId, int distance) {
+    public SectionRequest() {
+    }
+
+    public SectionRequest(Long upStationId, Long downStationId, int distance) {
         this.upStationId = upStationId;
         this.downStationId = downStationId;
         this.distance = distance;
@@ -16,23 +18,11 @@ public class SectionResponse {
         return upStationId;
     }
 
-    public void setUpStationId(Long upStationId) {
-        this.upStationId = upStationId;
-    }
-
     public Long getDownStationId() {
         return downStationId;
     }
 
-    public void setDownStationId(Long downStationId) {
-        this.downStationId = downStationId;
-    }
-
     public int getDistance() {
         return distance;
-    }
-
-    public void setDistance(int distance) {
-        this.distance = distance;
     }
 }
