+++ src/main/java/subway/station/StationController.java
@@ -3,10 +3,7 @@ package subway.station;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.web.bind.annotation.*;
-import subway.line.LineDao;
-import subway.line.SectionDao;
 
 import java.net.URI;
 import java.util.ArrayList;
